extract config dump to external tool
