+++ src/core/implementation/configuration.cpp
@@ -25,74 +25,15 @@
 #include <QDir>
 #include <QFile>
 #include <QFileInfo>
-#include <QJsonDocument>
-#include <QJsonObject>
-#include <QJsonValue>
 #include <QVariant>
 
 #include <system/system.hpp>
 
 #include "constants.hpp"
 
-namespace
-{
-    void writeTo(QJsonObject& obj, QStringList configPath, const QVariant& value)
-    {
-        const QString entryName = configPath.front();
 
-        if (configPath.size() == 1)
-        {
-            const QJsonValue json_value = QJsonValue::fromVariant(value);
-            assert(json_value.isUndefined() == false);
-            assert(json_value.isNull() == false);
-
-            obj[entryName] = json_value;
-        }
-        else
-        {
-            configPath.takeFirst();
-            QJsonValueRef jsonValue = obj[entryName];
-            assert(jsonValue.isNull() || jsonValue.isObject());
-
-            QJsonObject subObj = jsonValue.toObject();
-            writeTo(subObj, configPath, value);
-
-            jsonValue = subObj;
-        }
-    }
-
-    IConfigStorage::Content readNode(const QJsonObject& obj, const QString& entry_namespace = {})
-    {
-        IConfigStorage::Content content;
-
-        for(auto it = obj.constBegin(); it != obj.constEnd(); ++it)
-        {
-            const QString name = it.key();
-            const QJsonValue value = it.value();
-
-            assert(value.isArray() == false);
-            assert(value.isNull() == false);
-
-            const QString sub_namespace = entry_namespace + (entry_namespace.isEmpty()? name : QString("::%1").arg(name));
-
-            if (value.isObject())
-            {
-                const QJsonObject sub_obj = value.toObject();
-                const IConfigStorage::Content sub_content = readNode(sub_obj, sub_namespace);
-
-                content.insert(sub_content.cbegin(), sub_content.cend());
-            }
-            else
-                content.emplace(sub_namespace, value);
-        }
-
-        return content;
-    }
-}
-
-
-ConfigurationPrivate::ConfigurationPrivate(const QString& configFile):
-    m_configFile(configFile),
+ConfigurationPrivate::ConfigurationPrivate(IConfigStorage& configStorage):
+    m_configStorage(configStorage),
     m_entries(),
     m_dumpTimer(),
     m_watchers()
@@ -148,22 +89,8 @@ void ConfigurationPrivate::watchFor(const QString& key, const IConfiguration::Wa
 
 void ConfigurationPrivate::loadData()
 {
-    if (QFile::exists(m_configFile))
-    {
-        QFile configFile(m_configFile);
-        configFile.open(QIODevice::ReadOnly);
-
-        const QByteArray configFileContent = configFile.readAll();
-        const QJsonDocument jsonDoc = QJsonDocument::fromJson(configFileContent);
-
-        auto locked_config = m_entries.lock();
-        const QJsonObject configJsonObj = jsonDoc.object();
-        *locked_config = readNode(configJsonObj);
-    }
-    else
-    {
-        //load default data
-    }
+    auto locked_config = m_entries.lock();
+    *locked_config = m_configStorage.load();
 }
 
 
@@ -175,36 +102,16 @@ void ConfigurationPrivate::markDataDirty()
 
 void ConfigurationPrivate::saveData()
 {
-    const QFileInfo configPathInfo(m_configFile);
-    const QDir configDir(configPathInfo.absolutePath());
-
-    if (configDir.exists() == false)
-        configDir.mkpath(".");
-
     auto locked_config = m_entries.lock();
-
-    QJsonObject configurationObject;
-
-    for(const auto& [key, value]: *locked_config)
-    {
-        const QStringList entries = key.split("::");
-        writeTo(configurationObject, entries, value);
-    }
-
-    QJsonDocument jsonDoc(configurationObject);
-
-    QFile configFile(m_configFile);
-
-    configFile.open(QIODevice::WriteOnly);
-    configFile.write(jsonDoc.toJson());
+    m_configStorage.save(*locked_config);
 }
 
 
 ////////////////////////////////////////////////////////////////////////////////
 
 
-Configuration::Configuration(const QString& configFilePath):
-    d(new ConfigurationPrivate(configFilePath))
+Configuration::Configuration(IConfigStorage& configStorage):
+    d(new ConfigurationPrivate(configStorage))
 {
 
 }
