+++ tr/photo_broom_pl.ts
@@ -1080,7 +1080,7 @@ lub użyj menadżera pakietów swojej dystrybucji Linuksowej.
 <context>
     <name>main</name>
     <message>
-        <location filename="../src/main.cpp" line="+61"/>
+        <location filename="../src/main.cpp" line="+178"/>
         <source>Defines loging level. Possible options are: Debug, Info, Warning (default), Error</source>
         <translation>Definiuje poziom logowania. Dostępne opcje: Debug, Info, Warning (domyślnie), Error</translation>
     </message>
